requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.39
requests==2.28.1
numpy==1.23.4
pandas==1.5.1
flask-cors
psycopg2-binary
Werkzeug==2.0.3
mysql-connector-python